Make keychain storage unavailable on WASM
